+++ p2panda/src/node.rs
@@ -4,13 +4,12 @@ use std::fmt::Debug;
 use std::sync::Mutex;
 
 use futures_util::Stream;
-use p2panda_core::traits::ShortFormat;
 use p2panda_core::{Hash, Topic};
 use p2panda_net::connection_authoriser::ConnectionAuthoriser;
 use p2panda_net::iroh_endpoint::RelayUrl;
 use p2panda_net::{NetworkId, NodeId};
 use p2panda_spaces::manager::GLOBAL_GROUPS_CONTEXT_ID;
-use p2panda_spaces::{AuthGroupState, Config as SpacesConfig, GroupId, SpaceId, SpacesStoreState};
+use p2panda_spaces::{Config as SpacesConfig, GroupId, SpaceId, SpacesStoreState};
 use p2panda_store::groups::GroupsStore;
 use p2panda_store::spaces::{SpacesStore, SqliteSpacesStore};
 use p2panda_store::sqlite::{SqliteError, SqliteStore, SqliteStoreBuilder};
@@ -20,7 +19,6 @@ use p2panda_stream::hooks::ProcessorHooksList;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tokio::sync::broadcast;
-use tracing::debug;
 
 pub use crate::builder::NodeBuilder;
 use crate::credentials::Credentials;
@@ -28,12 +26,12 @@ use crate::forge::{Forge, OperationForge};
 use crate::network::{Network, NetworkConfig, NetworkError};
 use crate::operation::Extensions;
 use crate::spaces::types::{
-    AuthCapabilities, InnerSpace, InnerSpaceError, NoBody, SpacesManager, SpacesManagerError,
+    InnerSpace, InnerSpaceError, NoBody, SpacesManager, SpacesManagerError,
 };
 use crate::spaces::{
     AccessLevel, ActorId, ConnectionAuthoriserHook, DEFAULT_REPAIR_STRATEGY, Group, GroupError,
     KeyBundleTask, Member, MemberAssociationHook, MemberError, RepairTask, Space,
-    SpaceSubscription, actor_to_topic, group_log_id, member_log_id, spaces_manager, spaces_stream,
+    SpaceSubscription, actor_to_topic, member_log_id, spaces_manager, spaces_stream,
     to_initial_members,
 };
 use crate::streams::{
@@ -525,28 +523,6 @@ impl Node {
         let space_id = space_id.into();
 
         tx!(self.store, {
-            // Associate all group logs we have with the space topic, this handles the "first time
-            // subscription" case where we want to sync all groups logs up-front.
-            //
-            // TODO: This can be removed once we have a working orderer as then the repair task can
-            // be relied upon.
-            let y: AuthGroupState<AuthCapabilities> = self
-                .store
-                .get_groups_state_tx(Hash::digest(GLOBAL_GROUPS_CONTEXT_ID))
-                .await?
-                .unwrap_or_default();
-
-            for group_id in y.groups_global() {
-                debug!(
-                    group_id = group_id.fmt_short(),
-                    space_id = space_id.fmt_short(),
-                    "associate group log with space topic"
-                );
-                self.store
-                    .associate(&Topic::from(space_id), &self.id(), &group_log_id(group_id))
-                    .await?;
-            }
-
             // Associate the space topic with our own member / key bundle logs.
             self.store
                 .associate(&Topic::from(space_id), &self.id(), &member_log_id())