+++ p2panda/src/spaces/mod.rs
@@ -20,7 +20,7 @@ pub use p2panda_spaces::manager::ManagerError;
 pub use p2panda_spaces::{ActorId, GroupContext, GroupId, MemberId, SpaceContext, SpaceId};
 
 pub(crate) use authoriser::ConnectionAuthoriserHook;
-pub(crate) use forge::{group_log_id, member_log_id};
+pub(crate) use forge::member_log_id;
 pub use group::{
     AddGroupMemberError, Group, GroupError, GroupEvent, GroupFuture, RemoveGroupMemberError,
 };